Safari-609.1.11.1/LayoutTests/fast/viewport/viewport-101-expected.txt

ALERT: viewport size 400x440 scale 0.8 with limits [0.8, 5] and userScalable true
